Button.check_press input chanfged
check_press now expects a tuple, instead of seperate x,y.
(You'll always input the mouse.get_pos(), which comes as a tuple)
Also, couldn't you simply have x,y be the mouse.get_pos() in the first
place?
I mean, it's always going to want to check for the pos, of the mouse
anyway, right?